Update CONTRIBUTING.md
Dear @fellyph,

Thank you for the document. I have made some small changes but the document itself is a real addition to the repo. 

What do you think? Should we use this document as the main  readme.md? The contents of this document is a big improvement if I compare it with our current README.MD. Or should we just link to the CONTRIBUTING.md from the README.MD?

Kind regards,

Jack Kwakman